Add tonal debugging and gprs buffer reading
This adds tones to allow for debugging without the serial monitor. Lower
pitches represent earlier places in the code, pitches that are not on
the harmonics (e.g. 404) are used for errors.

In addition, this adds an extra check in read gprsSerial loop. If a bad
network connection splits the message, then this should try to
approximate when two messages are likely to actually be the same
message.